requirements.txt
insightface
opencv-python
openai-whisper
setuptools-rust
pyaudio 
speechrecognition
sqlalchemy
onnxruntime